refactor(server): update json response at ParkingMap service

+++ server/app/services/parking_map_service.rb
@@ -1,6 +1,9 @@
 class ParkingMapService
   # Create / recreate a parking map with a given number of rows and columns
   def self.reset_and_initialize_map(rows, cols)
+    # Validate minimum requirements
+    raise ArgumentError, "Minimum sizes for parking map are 4" if (rows*cols) < 4
+
     # Clear all existing data
     ActiveRecord::Base.transaction do
       ParkingTransaction.delete_all
@@ -11,7 +14,7 @@ class ParkingMapService
     end
 
     # Create parking slots
-    parking_slots =[]
+    parking_slots = []
     rows.times do |r|
       cols.times do |c|
         parking_slots << {
@@ -24,13 +27,13 @@ class ParkingMapService
     end
     ParkingSlot.insert_all(parking_slots)
 
-    # create 3 default random entry points
+    # Create 3 default random entry points
     generated_coords = Set.new
     3.times do
       loop do
         random_row = rand(rows)
         random_col = rand(cols)
-        # Ensure the randomize entry point coordinates are unique
+        # Ensure the randomized entry point coordinates are unique
         if generated_coords.add?([ random_row, random_col ])
           add_entry_point(random_row, random_col)
           break
@@ -38,7 +41,11 @@ class ParkingMapService
       end
     end
 
-    get_parking_map
+    {
+      success: true,
+      message: "Parking map created successfully",
+      data: get_parking_map
+    }
   end
 
   def self.add_entry_point(row, col)
@@ -58,7 +65,7 @@ class ParkingMapService
 
       new_parking_slot_distances = []
       ParkingSlot.all.each do |slot|
-        # calculate distance using manhattan
+        # Calculate distance using Manhattan distance
         distance = (slot.row - row).abs + (slot.col - col).abs
         new_parking_slot_distances << {
           parking_slot_id: slot.id,
@@ -66,9 +73,21 @@ class ParkingMapService
           distance: distance
         }
       end
-      ParkingSlotDistance.insert_all(new_parking_slot_distances)
+      ParkingSlotDistance.insert_all(new_parking_slot_distances) if new_parking_slot_distances.any?
     end
-    get_parking_map
+
+    {
+      success: true,
+      message: "Entry point added successfully",
+      data: {
+        entry_point: {
+          id: entry_point.id,
+          row: entry_point.row,
+          col: entry_point.col
+        },
+        updated_map: get_parking_map
+      }
+    }
   end
 
 
@@ -76,33 +95,40 @@ class ParkingMapService
     all_slots = ParkingSlot.all
     all_entry_points = EntryPoint.all
 
-    # find grid dimensions
+    # Find grid dimensions
     max_row = (all_slots.map(&:row) + all_entry_points.map(&:row)).max || -1
     max_col = (all_slots.map(&:col) + all_entry_points.map(&:col)).max || -1
 
     slots_data = all_slots.includes(parking_transactions: :vehicle).map do |slot|
-      occupied_info = {}
-
-      # add vehicle information if slot is occupied
-      if slot.is_occupied
-        transaction = slot.parking_transactions.parked.first
-        occupied_info = {
-          plate_number: transaction&.vehicle&.plate_number,
-          entry_time: transaction&.entry_time,
-          vehicle_type: transaction&.vehicle&.vehicle_type
-        }
-      end
-      {
+      slot_info = {
         id: slot.id,
         row: slot.row,
         col: slot.col,
         slot_type: slot.slot_type,
         is_occupied: slot.is_occupied
-      }.merge(occupied_info)
+      }
+
+      # Add vehicle information if slot is occupied
+      if slot.is_occupied
+        transaction = slot.parking_transactions.parked.first
+        if transaction
+          slot_info[:occupied_by] = {
+            plate_number: transaction.vehicle.plate_number,
+            vehicle_type: transaction.vehicle.vehicle_type,
+            entry_time: transaction.entry_time.iso8601
+          }
+        end
+      end
+
+      slot_info
     end
 
     entry_points_data = all_entry_points.map do |ep|
-      { id: ep.id, row: ep.row, col: ep.col }
+      {
+        id: ep.id,
+        row: ep.row,
+        col: ep.col
+      }
     end
 
     {
